Extract preferences related server handlers from main.ts to server/preferences/app.ts
