+++ packages/loot-core/src/types/server-handlers.d.ts
@@ -11,7 +11,6 @@ import {
   PayeeEntity,
 } from './models';
 import { OpenIdConfig } from './models/openid';
-import { GlobalPrefs, MetadataPrefs } from './prefs';
 // eslint-disable-next-line import/no-unresolved
 import { Query } from './query';
 import { EmptyObject } from './util';
@@ -117,14 +116,6 @@ export interface ServerHandlers {
 
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   query: (query: Query) => Promise<{ data: any; dependencies: string[] }>;
-
-  'save-global-prefs': (prefs) => Promise<'ok'>;
-
-  'load-global-prefs': () => Promise<GlobalPrefs>;
-
-  'save-prefs': (prefsToSet) => Promise<'ok'>;
-
-  'load-prefs': () => Promise<MetadataPrefs | null>;
 
   'sync-reset': () => Promise<{ error?: { reason: string; meta?: unknown } }>;
